Optimisation du calcul des coefficients des filtres de l'eq graphique.
