+++ source/3rdParty/make_fftw_unix.bash
@@ -1,5 +1,9 @@
 #!/bin/bash
 
-if [ ! -d ./build ]; then
+if [ ! -f ./build/lib/libfftw3f.a ]; then
     ./configure --enable-float --prefix=$PWD/build && make && make install
+fi
+
+if [ ! -f ./build/lib/libfftw3.a ]; then
+    ./configure --prefix=$PWD/build && make && make install
 fi
